backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
supabase==2.13.0
celery==5.3.6
redis==5.0.1
httpx>=0.27,<1.0
apify-client>=1.8.1
yt-dlp==2024.3.10
youtube-transcript-api==0.6.1
openai>=1.12.0
anthropic>=0.18.1
deepgram-sdk==3.2.1
elevenlabs>=1.0.0
python-dotenv==1.0.1
pytest==8.0.0
pytest-asyncio==0.23.5
loguru==0.7.2
google-generativeai==0.4.1
